Add project's expiry dates to profile of a PPL holder
