+++ pages/project/list/content/index.js
@@ -1,10 +1,5 @@
 module.exports = {
   searchText: 'Search by project title, licence holder or licence number',
-  diff: {
-    singular: 'Less than 1 month left',
-    plural: 'Less than {{diff}} months left',
-    standard: '{{diff}} months left'
-  },
   fields: {
     title: {
       label: 'Project title'
