+++ pages/project/list/views/index.jsx
@@ -1,13 +1,11 @@
 import React, { Fragment } from 'react';
 import { connect } from 'react-redux';
-import differenceInMonths from 'date-fns/difference_in_months';
-import format from 'date-fns/format';
-import classnames from 'classnames';
 import SearchBar from '../../../common/views/containers/search';
 import FilterSummary from '../../../common/views/containers/filter-summary';
 import DataTable from '../../../common/views/containers/datatable';
 import Snippet from '../../../common/views/containers/snippet';
 import Link from '../../../common/views/containers/link';
+import ExpiryDate from '../../../common/views/components/expiry-date';
 
 export const formatters = {
   licenceHolder: {
@@ -16,26 +14,6 @@ export const formatters = {
   expiryDate: {
     format: date => <ExpiryDate date={date}/>
   }
-};
-
-const ExpiryDate = ({date}) => {
-  const diff = differenceInMonths(date, new Date());
-  const urgent = diff < 3;
-  let contentKey = 'diff.standard';
-  if (urgent) {
-    contentKey = diff === 0 ? 'diff.singular' : 'diff.plural';
-  }
-
-  return <Fragment>
-    { format(date, 'DD MMMM YYYY') }
-    {
-      diff < 12 && (
-        <span className={classnames('notice', { urgent })}>
-          <Snippet diff={diff + 1}>{contentKey}</Snippet>
-        </span>
-      )
-    }
-  </Fragment>;
 };
 
 const Projects = ({
